hotwheels-monitor/requirements.txt
requests
pyyaml
beautifulsoup4
PyGithub
python-dotenv
twilio
python-telegram-bot
